Split hook/mod.rs into several files
