+++ libimagstore/src/hook/read.rs
@@ -1,0 +1,13 @@
+use storeid::StoreId;
+use store::FileLockEntry;
+use hook::accessor::HookDataAccessorProvider;
+use hook::result::HookResult;
+use hook::Hook;
+
+pub trait PreReadHook : Hook {
+    fn pre_read(&self, &StoreId) -> HookResult<()>;
+}
+
+pub trait PostReadHook : Hook + HookDataAccessorProvider {
+}
+
