+++ libimagstore/src/hook/result.rs
@@ -1,0 +1,3 @@
+use hook::error::HookError;
+
+pub type HookResult<T> = Result<T, HookError>;
